json web token function added

+++ index.js
@@ -11,6 +11,8 @@ const port = process.env.PORT || 5000;
 app.use(cors());
 app.use(express.json());
 
+const secretJWT = process.env.REACT_APP_JWT_SECRET_ACCESS_TOKEN;
+
 // database config
 const uri = process.env.DB_URI;
 const client = new MongoClient(uri, {
@@ -18,6 +20,26 @@ const client = new MongoClient(uri, {
   useUnifiedTopology: true,
   serverApi: ServerApiVersion.v1,
 });
+
+// token verify functuon
+const verifyTOken = (req, res, next) => {
+  const bearerToken = req.headers.authorization;
+
+  if (!bearerToken) {
+    return res.status(401).sent({ message: "unauthorazin access" });
+  } else {
+    const token = bearerToken.split(" ")[1];
+
+    jwt.verify(token, secretJWT, (err, decoded) => {
+      if (err) {
+        console.log(err);
+        return res.status(401).sent({ message: "unauthorazin access" });
+      }
+      req.decoded = decoded;
+    });
+  }
+  next();
+};
 
 const run = async () => {
   const serviceCollection = client.db("travel_point").collection("services");
@@ -27,55 +49,23 @@ const run = async () => {
     // JSON WEB TOKEN
     app.post("/jwt", (req, res) => {
       const user = req.body;
-      const token = jwt.sign(
-        user,
-        process.env.REACT_APP_JWT_SECRET_ACCESS_TOKEN,
-        {
-          expiresIn: "1d",
-        }
-      );
+      const token = jwt.sign(user, secretJWT, {
+        expiresIn: "1d",
+      });
       res.send({ token });
     });
 
-    // token verify functuin
-    const verifyTOken = (req, res, next) => {
-      const bearerToken = req.headers.authorization;
+    // home API =======================================================>
+    app.get("/home", async (req, res) => {
+      const query = {};
+      const cursor = serviceCollection.find(query).sort({ publish: -1 });
+      const result = await cursor.limit(3).toArray();
 
-      console.log(bearerToken);
-
-      if (!bearerToken) {
-        return res.status(401).sent({ message: "unauthorazin access" });
-      } else {
-        const token = bearerToken.split(" ")[1];
-
-        const secret = process.env.REACT_APP_JWT_SECRET_ACCESS_TOKEN;
-
-        console.log(secret);
-        console.log(token);
-
-        jwt.verify(token, secret, (err, decoded) => {
-          if (err) {
-            console.log(err);
-            return res.status(401).sent({ message: "unauthorazin access" });
-          }
-          req.decoded = decoded;
-        });
-      }
-
-      // next();
-    };
-
-    app.post("/service", async (req, res) => {
-      const service = req.body;
-
-      // const decoded = req.decoded;
-
-      // console.log(decoded);
-
-      const result = await serviceCollection.insertOne(service);
       res.send(result);
     });
+    // home API ==========================================================>
 
+    // review API ==================================>
     app.post("/review", async (req, res) => {
       const review = req.body;
       const result = await reviewCollection.insertOne(review);
@@ -92,60 +82,87 @@ const run = async () => {
       res.send(result);
     });
 
-    app.get("/my-reviews", async (req, res) => {
+    app.get("/my-reviews", verifyTOken, async (req, res) => {
       const userEmail = req.query.email;
+      const decoded = req.decoded;
 
-      const query = { email: userEmail };
-      const cursor = reviewCollection.find(query).sort({ postTime: -1 });
-      const result = await cursor.toArray();
+      if (decoded !== decoded) {
+        res.status(403).send({ message: "unauthorized access" });
+      } else {
+        const query = { email: userEmail };
+        const cursor = reviewCollection.find(query).sort({ postTime: -1 });
+        const result = await cursor.toArray();
 
-      res.send(result);
-    });
-
-    app.put("/my-reviews", async (req, res) => {
-      const id = req.query._id;
-      const filter = { _id: ObjectId(id) };
-      const review = req.body;
-      const option = { upsert: true };
-      const updateReview = {
-        $set: {
-          review: review.review,
-          userReating: review.userReating,
-        },
-      };
-      const result = await reviewCollection.updateOne(
-        filter,
-        updateReview,
-        option
-      );
-      res.send(result);
-    });
-
-    app.delete("/my-reviews", async (req, res) => {
-      const id = req.query._id;
-
-      const query = { _id: ObjectId(id) };
-      const result = await reviewCollection.deleteOne(query);
-
-      if (result.deletedCount === 1) {
         res.send(result);
       }
     });
 
-    app.get("/home", async (req, res) => {
-      const query = {};
-      const cursor = serviceCollection.find(query).sort({ publish: -1 });
-      const result = await cursor.limit(3).toArray();
+    app.put("/my-reviews", verifyTOken, async (req, res) => {
+      const id = req.query._id;
+      const userEmail = req.query.email;
+      const decoded = req.decoded.email;
 
-      res.send(result);
+      console.log(userEmail, id, decoded);
+      if (decoded !== userEmail) {
+        res.status(403).send({ message: "unauthorized access" });
+      } else {
+        const filter = { _id: ObjectId(id) };
+        const review = req.body;
+        const option = { upsert: true };
+        const updateReview = {
+          $set: {
+            review: review.review,
+            userReating: review.userReating,
+          },
+        };
+        const result = await reviewCollection.updateOne(
+          filter,
+          updateReview,
+          option
+        );
+        res.send(result);
+      }
     });
 
+    app.delete("/my-reviews", verifyTOken, async (req, res) => {
+      const id = req.query._id;
+      const userEmail = req.query.email;
+      const decoded = req.decoded;
+
+      if (decoded !== userEmail) {
+        res.status(403).send({ message: "unauthorized access" });
+      } else {
+        const query = { _id: ObjectId(id) };
+        const result = await reviewCollection.deleteOne(query);
+
+        if (result.deletedCount === 1) {
+          res.send(result);
+        }
+      }
+    });
+    // review API  ==============================================>
+
+    // service API ==========================================================>
     app.get("/services", async (req, res) => {
       const query = {};
       const cursor = serviceCollection.find(query).sort({ publish: -1 });
       const result = await cursor.toArray();
 
       res.send(result);
+    });
+
+    app.post("/service", verifyTOken, async (req, res) => {
+      const service = req.body;
+
+      const decoded = req.decoded.email;
+      const userEmail = service.authorEmail;
+
+      if (decoded !== userEmail) {
+        res.status(403).send({ message: "unauthorized access" });
+      } else {
+        const result = await serviceCollection.insertOne(service);
+        res.send(result);
+      }
     });
 
     app.get("/services/:id", async (req, res) => {
@@ -159,26 +176,40 @@ const run = async () => {
       res.send(result);
     });
 
-    app.get("/my-services", async (req, res) => {
+    app.get("/my-services", verifyTOken, async (req, res) => {
       const userEmail = req.query.email;
 
-      const query = { authorEmail: userEmail };
-      const cursor = serviceCollection.find(query).sort({ postTime: -1 });
-      const result = await cursor.toArray();
+      const decoded = req.decoded.email;
 
-      res.send(result);
-    });
+      if (decoded !== userEmail) {
+        res.status(403).send({ message: "unauthorized access" });
+      } else {
+        const query = { authorEmail: userEmail };
+        const cursor = serviceCollection.find(query).sort({ postTime: -1 });
+        const result = await cursor.toArray();
 
-    app.delete("/my-services", async (req, res) => {
-      const id = req.query._id;
-
-      const query = { _id: ObjectId(id) };
-      const result = await serviceCollection.deleteOne(query);
-
-      if (result.deletedCount === 1) {
         res.send(result);
       }
     });
+
+    app.delete("/my-services", verifyTOken, async (req, res) => {
+      const id = req.query._id;
+      const email = req.query.email;
+      const decoded = req.decoded;
+      console.log(email, decoded.email);
+
+      if (decoded !== email) {
+        res.status(403).send({ message: "unauthorized access" });
+      } else {
+        const query = { _id: ObjectId(id) };
+        const result = await serviceCollection.deleteOne(query);
+
+        if (result.deletedCount === 1) {
+          res.send(result);
+        }
+      }
+    });
+    //  service API ======================================================>
   } finally {
   }
 };
